ssu-repos: Define adaptation in both rnd and release mode
ssu-vendor-data-sailfishosorg doesn't define adaptation properly.
It was assumed that will be preferred over the previous definition of
the base ssu-vendor-data no matter if release or rnd mode if just defined in the `all` section.
However since the base config defines adaptation in each release mode those section are preferred and the path of the adaptation repo is wrong.

[ssu-repos] Define adaptation in both rnd and release mode. Fixes SailfishOS-SonyXperia/bugs#49

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>